pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi_agent_app"
version = "0.1.0"
description = "Enterprise-grade LangGraph Multi-Agent boilerplate"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-nvidia-ai-endpoints>=0.3",
    "langchain-openai>=0.2",
    "langchain-google-genai>=2.0",
    "langchain-qdrant>=0.2",
    "langchain-huggingface>=0.1",
    "langgraph>=0.2",
    "langgraph-cli[inmem]>=0.1.55",
    "qdrant-client>=1.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
